+++ tests/test_service_v2.py
@@ -8,7 +8,8 @@ from unittest.mock import patch
 
 import pyodata.v2.model
 import pyodata.v2.service
-from pyodata.exceptions import PyODataException, HttpError, ExpressionError, ProgramError
+from pyodata.exceptions import PyODataException, HttpError, ExpressionError, ProgramError, PyODataModelError
+from pyodata.v2 import model
 from pyodata.v2.service import EntityKey, EntityProxy, GetEntitySetFilter, ODataHttpResponse, HTTP_CODE_OK
 
 from tests.conftest import assert_request_contains_header, contents_of_fixtures_file
@@ -22,6 +23,13 @@ def service(schema):
     """Service fixture"""
     assert schema.namespaces   # this is pythonic way how to check > 0
     return pyodata.v2.service.Service(URL_ROOT, schema, requests)
+
+
+@pytest.fixture
+def service_retain_null(schema):
+    """Service fixture which keeps null values as such"""
+    assert schema.namespaces
+    return pyodata.v2.service.Service(URL_ROOT, schema, requests, model.Config(retain_null=True))
 
 
 @responses.activate
@@ -286,7 +294,7 @@ def test_entity_key_complex(service):
 
     entity_key = {
         'Sensor': 'sensor1',
-        'Date': datetime.datetime(2017, 12, 24, 18, 0)
+        'Date': datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc)
     }
     key_properties = set(entity_key.keys())
 
@@ -345,7 +353,7 @@ def test_entity_key_complex_valid(service):
 
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
-        Sensor='sensor1', Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Sensor='sensor1', Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     assert key.to_key_string() == "(Sensor='sensor1',Date=datetime'2017-12-24T18:00:00')"
 
@@ -616,13 +624,13 @@ def test_update_entity(service):
 
     request = service.entity_sets.TemperatureMeasurements.update_entity(
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     assert isinstance(request, pyodata.v2.service.EntityModifyRequest)
 
     request.set(Value=34.0)
     # Tests if update entity correctly calls 'to_json' method
-    request.set(Date=datetime.datetime(2017, 12, 24, 19, 0))
+    request.set(Date=datetime.datetime(2017, 12, 24, 19, 0, tzinfo=datetime.timezone.utc))
 
     assert request._values['Value'] == '3.400000E+01'
     assert request._values['Date'] == '/Date(1514142000000)/'
@@ -684,7 +692,7 @@ def test_update_entity_with_entity_key(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.update_entity(key)
     assert query.get_path() == "TemperatureMeasurements(Sensor='sensor1',Date=datetime'2017-12-24T18:00:00')"
@@ -699,7 +707,7 @@ def test_update_entity_with_put_method_specified(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.update_entity(key, method="PUT")
     assert query.get_method() == "PUT"
@@ -714,7 +722,7 @@ def test_update_entity_with_patch_method_specified(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.update_entity(key, method="PATCH")
     assert query.get_method() == "PATCH"
@@ -728,7 +736,7 @@ def test_update_entity_with_merge_method_specified(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.update_entity(key, method='merge')
     assert query.get_method() == 'MERGE'
@@ -743,7 +751,7 @@ def test_update_entity_with_no_method_specified(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.update_entity(key)
     assert query.get_method() == "PATCH"
@@ -758,7 +766,7 @@ def test_update_entity_with_service_config_set_to_put(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     service.config['http']['update_method'] = "PUT"
     query = service.entity_sets.TemperatureMeasurements.update_entity(key)
@@ -774,7 +782,7 @@ def test_update_entity_with_wrong_method_specified(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     with pytest.raises(ValueError) as caught_ex:
         service.entity_sets.TemperatureMeasurements.update_entity(key, method='DELETE')
@@ -790,7 +798,7 @@ def test_get_entity_with_entity_key_and_other_params(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.update_entity(key=key, Foo='Bar')
     assert query.get_path() == "TemperatureMeasurements(Sensor='sensor1',Date=datetime'2017-12-24T18:00:00')"
@@ -807,7 +815,7 @@ def test_get_entity_with_custom_headers(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.get_entity(key)
     query.add_headers({"X-Foo": "bar"})
@@ -819,7 +827,7 @@ def test_update_entities_with_custom_headers(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.update_entity(key)
     query.add_headers({"X-Foo": "bar"})
@@ -884,6 +892,88 @@ def test_get_entities(service):
     assert empls[0].ID == 669
     assert empls[0].NameFirst == 'Yennefer'
     assert empls[0].NameLast == 'De Vengerberg'
+
+
+@responses.activate
+def test_get_null_value_from_null_preserving_service(service_retain_null):
+    """Get entity with missing property value as None type"""
+
+    # pylint: disable=redefined-outer-name
+
+    responses.add(
+        responses.GET,
+        f"{service_retain_null.url}/Employees",
+        json={'d': {
+            'results': [
+                {
+                    'ID': 1337,
+                    'NameFirst': 'Neo',
+                    'NameLast': None
+                }
+            ]
+        }},
+        status=200)
+
+    request = service_retain_null.entity_sets.Employees.get_entities()
+
+    the_ones = request.execute()
+    assert the_ones[0].ID == 1337
+    assert the_ones[0].NameFirst == 'Neo'
+    assert the_ones[0].NameLast is None
+
+
+@responses.activate
+def test_get_null_value_from_non_null_preserving_service(service):
+    """Get entity with missing property value as default type"""
+
+    # pylint: disable=redefined-outer-name
+
+    responses.add(
+        responses.GET,
+        f"{service.url}/Employees",
+        json={'d': {
+            'results': [
+                {
+                    'ID': 1337,
+                    'NameFirst': 'Neo',
+                    'NameLast': None
+                }
+            ]
+        }},
+        status=200)
+
+    request = service.entity_sets.Employees.get_entities()
+
+    the_ones = request.execute()
+    assert the_ones[0].ID == 1337
+    assert the_ones[0].NameFirst == 'Neo'
+    assert the_ones[0].NameLast == ''
+
+
+@responses.activate
+def test_get_non_nullable_value(service_retain_null):
+    """Get error when receiving a null value for a non-nullable property"""
+
+    # pylint: disable=redefined-outer-name
+
+    responses.add(
+        responses.GET,
+        f"{service_retain_null.url}/Employees",
+        json={'d': {
+            'results': [
+                {
+                    'ID': None,
+                    'NameFirst': 'Neo',
+                }
+            ]
+        }},
+        status=200)
+
+    with pytest.raises(PyODataException) as e_info:
+        service_retain_null.entity_sets.Employees.get_entities().execute()
+
+    assert str(e_info.value) == 'Value of non-nullable Property ID is null'
+
 
 @responses.activate
 def test_navigation_multi(service):
@@ -1380,7 +1470,7 @@ def test_batch_request(service):
 
     temp_request = service.entity_sets.TemperatureMeasurements.update_entity(
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0)).set(Value=34.0)
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc)).set(Value=34.0)
 
     batch.add_request(employee_request)
 
@@ -1480,7 +1570,7 @@ def test_get_entity_with_entity_key(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.get_entity(key)
     assert query.get_path() == "TemperatureMeasurements(Sensor='sensor1',Date=datetime'2017-12-24T18:00:00')"
@@ -1494,7 +1584,7 @@ def test_get_entity_with_entity_key_and_other_params(service):
     key = EntityKey(
         service.schema.entity_type('TemperatureMeasurement'),
         Sensor='sensor1',
-        Date=datetime.datetime(2017, 12, 24, 18, 0))
+        Date=datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc))
 
     query = service.entity_sets.TemperatureMeasurements.get_entity(key=key, Foo='Bar')
     assert query.get_path() == "TemperatureMeasurements(Sensor='sensor1',Date=datetime'2017-12-24T18:00:00')"
@@ -2259,9 +2349,38 @@ def test_count_with_chained_filters(service):
 
 
 @responses.activate
-def test_create_entity_with_datetime(service):
+def test_create_entity_with_utc_datetime(service):
+    """Basic test on creating entity with an UTC datetime object"""
+
+    # pylint: disable=redefined-outer-name
+
+    responses.add(
+        responses.POST,
+        f"{service.url}/TemperatureMeasurements",
+        headers={'Content-type': 'application/json'},
+        json={'d': {
+            'Sensor': 'Sensor1',
+            'Date': '/Date(1514138400000)/',
+            'Value': '34.0d'
+        }},
+        status=201)
+
+    request = service.entity_sets.TemperatureMeasurements.create_entity().set(**{
+        'Sensor': 'Sensor1',
+        'Date': datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc),
+        'Value': 34.0
+    })
+
+    assert request._values['Date'] == '/Date(1514138400000)/'
+
+    result = request.execute()
+    assert result.Date == datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc)
+
+
+@responses.activate
+def test_create_entity_with_non_utc_datetime(service):
     """
-        Basic test on creating entity with datetime
+        Basic test on creating entity with an non-UTC datetime object
         Also tzinfo is set to simulate user passing datetime object with different timezone than UTC
     """
 
@@ -2294,18 +2413,48 @@ def test_create_entity_with_datetime(service):
         }},
         status=201)
 
+    with pytest.raises(PyODataModelError) as e_info:
+        # Offset -18000 sec is for America/Chicago (CDT) timezone
+        service.entity_sets.TemperatureMeasurements.create_entity().set(**{
+            'Sensor': 'Sensor1',
+            'Date': datetime.datetime(2017, 12, 24, 18, 0, tzinfo=MyUTCOffsetTimezone(-18000)),
+            'Value': 34.0
+        })
 
-    # Offset -18000 sec is for America/Chicago (CDT) timezone
-    request = service.entity_sets.TemperatureMeasurements.create_entity().set(**{
-        'Sensor': 'Sensor1',
-        'Date': datetime.datetime(2017, 12, 24, 18, 0, tzinfo=MyUTCOffsetTimezone(-18000)),
-        'Value': 34.0
-    })
 
-    assert request._values['Date'] == '/Date(1514138400000)/'
+@responses.activate
+def test_create_entity_with_naive_datetime(service):
+    """Preventing creation/usage of an entity with an unaware datetime object"""
 
-    result = request.execute()
-    assert result.Date == datetime.datetime(2017, 12, 24, 18, 0, tzinfo=datetime.timezone.utc)
+    with pytest.raises(PyODataModelError) as e_info:
+        service.entity_sets.TemperatureMeasurements.create_entity().set(**{
+            'Sensor': 'Sensor1',
+            'Date': datetime.datetime(2017, 12, 24, 18, 0),
+            'Value': 34.0
+        })
+    assert str(e_info.value).startswith('Edm.DateTime accepts only UTC')
+
+
+@responses.activate
+def test_null_datetime(service):
+    """Test default value of DateTime. Default value gets inserted when a property is null"""
+
+    responses.add(
+        responses.GET,
+        f"{service.url}/TemperatureMeasurements",
+        headers={'Content-type': 'application/json'},
+        json={'d': {
+            'results': [
+                {
+                    'Date': None,
+                }
+            ]
+        }},
+        status=200)
+
+    result = service.entity_sets.TemperatureMeasurements.get_entities().execute()
+
+    assert result[0].Date == datetime.datetime(1753, 1, 1, 0, 0, 0, tzinfo=datetime.timezone.utc)
 
 
 @responses.activate
@@ -2327,7 +2476,7 @@ def test_parsing_of_datetime_before_unix_time(service):
 
     request = service.entity_sets.TemperatureMeasurements.create_entity().set(**{
         'Sensor': 'Sensor1',
-        'Date': datetime.datetime(1945, 5, 8, 19, 0),
+        'Date': datetime.datetime(1945, 5, 8, 19, 0, tzinfo=datetime.timezone.utc),
         'Value': 34.0
     })
 
@@ -2335,6 +2484,40 @@ def test_parsing_of_datetime_before_unix_time(service):
 
     result = request.execute()
     assert result.Date == datetime.datetime(1945, 5, 8, 19, 0, tzinfo=datetime.timezone.utc)
+
+
+@responses.activate
+@pytest.mark.parametrize("json_input,expected", [
+    ('/Date(981173106000+0001)/', datetime.datetime(2001, 2, 3, 4, 5, 6,
+                                                    tzinfo=datetime.timezone(datetime.timedelta(minutes=1)))),
+    ('/Date(981173106000-0001)/', datetime.datetime(2001, 2, 3, 4, 5, 6,
+                                                    tzinfo=datetime.timezone(-datetime.timedelta(minutes=1)))),
+    (None, datetime.datetime(1753, 1, 1, 0, 0, 0, tzinfo=datetime.timezone.utc)),
+])
+def test_parsing_of_datetimeoffset(service, json_input, expected):
+    """Test DateTimeOffset handling."""
+
+    # pylint: disable=redefined-outer-name
+
+    responses.add(
+        responses.GET,
+        f"{service.url}/TemperatureMeasurements",
+        headers={'Content-type': 'application/json'},
+        json={'d': {
+            'results': [
+                {
+                    'Sensor': 'Sensor1',
+                    'Date': '/Date(-981173106000)/',
+                    'DateTimeWithOffset': json_input,
+                    'Value': '34.0d'
+                }
+            ]
+        }},
+        status=200)
+
+    result = service.entity_sets.TemperatureMeasurements.get_entities().execute()
+
+    assert result[0].DateTimeWithOffset == expected
 
 
 @responses.activate