+++ CHANGELOG.md
@@ -6,6 +6,31 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.9.0]
+
+### Added
+- Prevent substitution of missing, nullable values - Reto Schneider
+
+### Fixed
+- Fix Increased robustness when schema with empty properties is returned - Andre Gensler
+- Use valid default value for Edm.DateTimeOffset - Reto Schneider
+
+### Changed
+- Adjusted Edm.DateTime default value to January 1, 1753 A.D - Reto Schneider
+
+## [1.8.0]
+
+### Fixed
+- Fix Edm.Binary literal representation - Daniel Balko
+- Datetime support for Edm.DateTimeOffset - Reto Schneider
+- Disallow creation of non-UTC Edm.DateTime - Reto Schneider
+
+### Changed
+- Split properties schema.entity_types/complex_types and their generated Collections - Petr Hanak
+
+### Removed
+- Python 3.6 (after its EOL) is no longer supported by pyodata. Python 3.7 is now minimal supported version. - Petr Hanak
+
 ## [1.7.1]
 
 ### Fixed
@@ -73,7 +98,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - support deleting Entities - Martin Miksik
 
 ### Changed
-- Emd.Int64 literals do no need to have the suffix L - Jakub Filak
+- Edm.Int64 literals do no need to have the suffix L - Jakub Filak
 - more user friendly Function call errors - Jakub Filak
 
 ### Fixed
@@ -81,7 +106,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - correctly serialize DateTime values to JSON in create/update methods - Martin Miksik
 - remove timezone info from DateTime URL literals - Martin Miksik
 
-# # [1.2.3]
+## [1.2.3]
 
 ### Added
 - add support for whitelisted and custom namespaces - Martin Miksik
@@ -118,7 +143,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.1.1 - First PIP package release
 
-[Unreleased]: https://github.com/SAP/python-pyodata/compare/1.7.1...HEAD
+[Unreleased]: https://github.com/SAP/python-pyodata/compare/1.9.0...HEAD
+[1.9.0]: https://github.com/SAP/python-pyodata/compare/1.8.0...1.9.0
+[1.8.0]: https://github.com/SAP/python-pyodata/compare/1.7.1...1.8.0
 [1.7.1]: https://github.com/SAP/python-pyodata/compare/1.7.0...1.7.1
 [1.7.0]: https://github.com/SAP/python-pyodata/compare/1.6.0...1.7.0
 [1.6.0]: https://github.com/SAP/python-pyodata/compare/1.5.0...1.6.0